+++ montrek/baseclasses/tests/test_views.py
@@ -1,10 +1,10 @@
+from baseclasses.tests.mocks import MockRepository
 from user.tests.factories.montrek_user_factories import MontrekUserFactory
 from django.contrib.auth.models import AnonymousUser
 from django.test import TestCase, RequestFactory
 from django.contrib.sessions.middleware import SessionMiddleware
 from django.contrib.messages.middleware import MessageMiddleware
 from django.contrib.messages import get_messages
-from dataclasses import dataclass
 from baseclasses.views import MontrekViewMixin
 from baseclasses.views import MontrekListView
 from baseclasses.dataclasses.montrek_message import (
@@ -14,64 +14,6 @@ from baseclasses.dataclasses.montrek_message import (
 from baseclasses.pages import MontrekPage
 from reporting.managers.montrek_table_manager import MontrekTableManager
 from reporting.dataclasses import table_elements as te
-
-
-class MockQuerySet:
-    def __init__(self, *args):
-        self.items = args
-
-    def __iter__(self):
-        return iter(self.items)
-
-    def __len__(self):
-        return len(self.items)
-
-    def __getitem__(self, index):
-        return self.items[index]
-
-    def __eq__(self, other):
-        if isinstance(other, list):
-            return list(self.items) == other
-        return NotImplemented
-
-    def all(self):
-        return self.items
-
-    def count(self):
-        return len(self.items)
-
-
-@dataclass
-class MockData:
-    field: str
-    value: int
-
-
-@dataclass
-class MockField:
-    name: str
-
-
-class MockRepository:
-    def __init__(self, session_data):
-        self.session_data = session_data
-        self.messages = []
-        self.annotations = {}
-
-    def receive(self):
-        return MockQuerySet(
-            MockData("item1", 1), MockData("item2", 2), MockData("item3", 3)
-        )  # Dummy data for testing
-
-    def std_satellite_fields(self):
-        return [
-            MockField("item1"),
-            MockField("item2"),
-            MockField("item3"),
-        ]  # Dummy data for testing
-
-    def get_all_fields(self):
-        return ["item1", "item2", "item3"]
 
 
 class MockRequester:
